feat(ui): refactor logic into controller
